+++ src/app/Types/cards.ts
@@ -1,0 +1,9 @@
+export type CardProp = {
+    title: string;
+    description: React.ReactNode | string;
+    author?: {
+        username: string;
+    };
+    shouldRender?: (loggedIn?: boolean) => boolean;
+    loggedIn?: boolean;
+};
